refactor: move reporter config logic to own file
