+++ src/test-results.ts
@@ -3,6 +3,7 @@ import { TestCaseData, TestsPerSpecFile, TestCaseError} from "./TestsPerSpecFile
 import Color from "../color-text/Color";
 import defineConfig from "../playwright.config";
 import { filterOutDuplicateFailedTests } from "./filtering-tests";
+import { getReporterOptions } from "./reporter-configuration";
 
 export interface StatusCounter {
   passed: number;
@@ -30,23 +31,6 @@ export const howToReadTestResults = () => {
   log(lineBreak);
 };
 
-export const isIndentedListReporter = (configuredReporters:  any[] | any): [boolean, number] | boolean => {
-  let isIndentedListReporter: [boolean, number] | boolean = false;
-  configuredReporters.forEach((reporter, index) => {
-    // or ./src/.. is for local testing before publishing the plugin in npm
-    if(reporter[0] === "indent-list-reporter" || reporter[0] === "./src/indent-list-reporter.ts") {
-      isIndentedListReporter = [true, index];
-    }
-  });
-  return isIndentedListReporter
-}
-
-export const getReporterOptions = (reporters: any[] | any): any => {
-  const reporterIndex = isIndentedListReporter(reporters);
-    if(isIndentedListReporter(reporters)) {
-      return reporters[reporterIndex[1]][1];
-    }
-}
 export const logSpecFileName = (specFileName: string) => {
   const reporterOptions = getReporterOptions(defineConfig.reporter);
   const specFileNameColor = reporterOptions?.baseColors?.specFileNameColor ? reporterOptions.baseColors.specFileNameColor : undefined;
